output break statements
The v8-native-prototype expects the break statements to give a break
depth, not a label name. The break depth starts at 0 and counts up for
each block or loop, going outward.

The WebAssembly spec uses label blocks instead, also starting at 0 and
counting up going outward. Blocks and loops are skipped, however. So we
keep a "v8" depth for each block or loop added, and introduce a new
block for every label found.