include filename in error output
The test infrastructure currently autogenerates source files from the
test .txt files, so this change modifies it to generate a temporary
directory instead, which is populated with test files with the same
path.

It then runs using the temporary directory as cwd so the error output
produces the correct file.