start of tracking continuations in expressions
This is needed to properly handle typechecking implicit returns and
return statements.

The type of a return statement should be void. For example, this should
be a type error even though the return type is f32:

  (f32.neg (return (f32.const 1))

But an implicit return should be legal:

  (func (result i32) (i32.const 10))

And an explicit return should be legal as well:

  (func (result i32) (return (i32.const 10)))

This last case requires knowing that this statement was a return
statement. That is not sufficient however, as this should be legal as
well:

  (func (result i32)
    (if (i32.const 1)
      (return (i32.const 2))
      (return (i32.const 3))))

This is easily resolved if we propagate continuations.