support exporting function with multiple names
This is enough to handle spec tests, but it won't actually export the
function with any name except the first. The v8-native-prototype
currently only allows a function to be exported with one name. We could
work around that by introducing a new function that forwards to the
other one, but this isn't implemented yet.