Rename assert_eq to assert_return
This matches recent changes in the spec repo. I left assert_eq as an
alias for assert_return to accomodate existing tests for now (although
the semantics for FP compares are subtly different), but renamed the
enums and event names to assert_return. I also left the d8
and spec tests alone for now.