update v8-native-prototype w/ new binary format
The new binary format has optional sections, which are marked with a
byte. The names and segment data must come after all the sections.

The other major change to the format is that function bodies are now
inlined with the function header. This makes it much easier to append
the assert_* functions to the module after the fact; we can place the
function section last, and write the assertion functions directly into
the buffer.

The only tricky part is that the function count is now stored as leb128,
which means that we may need to shift the function section down when the
function count increases.