make parser grammer lists left-recursive
expr_list is tricky, it can be made left-recursive, but func_fields has
to be modified to do so:

    func_fields :
        /* empty */
      | expr expr_list
      | LPAR PARAM value_type_list RPAR func_fields
      ...

This is because when expr_list is left-recursive, the rule to make an
empty expr_list is on the left, so there is a shift/reduce conflict:

    ( FUNC . (

In this case, it is impossible to whether this is the start of a
(currently empty) expr_list or the start of a parameter, result or
local.

The fix above prevents there from being an empty reduction, so the LPAR
can always be shifted.