[fuzz] fix assert with (memory 1) in gcc
__builtin_clz(x) is undefined when x is zero in gcc, but seems to often
return a value > 32. Since log_two_u32 is implemented as:

    sizeof(unsigned int) * 8 - __builtin_clz(x - 1)

This would return:

    32 - <something greater than 32>

Which as a uint32_t would wrap to some very large number. This would
then hit the assertion when writing a u8 value > 255.

It's interesting that this never seems to happen with clang, it always
returns 32 which produces log_two_u32(1) => 0.