write spec files as .json with .wasm files
sexpr-wasm --spec used to write a .js file that could be run by d8. This
isn't very convenient in general; and it won't be easy to use with
wasm-interp either.

Now it generates a json file which references a collection of .wasm
files. test/spec.js has also been modified to read this json file.