better error handling between squirrel + wasm
A wasm import function can now raise an error, which will propagate back
through the interpreter. This error can not be caught in wasm, but can
be caught in squirrel.

The other major change here is to prevent calling the error handler
multiple times. We store the currently propagating error as soon as we
see it. If we unwind and see it again, then we don't call the error
handler. It's a nasty hack, but it seems to work.