fix bugs in if_else code generation
The size of the value stack must be the same after each branch of an
if_else op. If the true branch has a value and the false branch doesn't,
then we'll have kept the value on the stack on the true branch, but
won't have a value on the false branch. To fix this, we write a NOP at
the end of the true branch, and if the value needs to be discarded, we
update it to a DISCARD.