fix bug when displaying source lines for errors
When scanning for the the file offset of a source line, when the source
is a file, it reuses the same code for scanning a memory buffer. If the
line is not found in the buffer, it returns WASM_ERROR, which is
expected. But it also needs to update the currently found line, so we
can keep scanning.