defer decoding binary modules in assert_invalid
If we decode the binary module when parsing, then any errors will be
displayed too early. Instead, we store the binary data, ready to be
decoded.

This is only done for assert_invalid, where we assume that the module
will not actually be used. Otherwise, it is more convenient to always
have a WasmModule, so we decode directly while parsing the AST.