WasmFuncDeclaration does not always set `sig`
Prior to this change, the AST parser would always set `sig`, even if the
function had no explicit signature (i.e. it used a function type
variable instead).

This is OK, but it makes the flag confusing.
`WASM_FUNC_DECLARATION_FLAG_HAS_SIGNATURE` will be clear and
`wasm_decl_has_signature` will return false, but `sig` will be valid.

This change makes it so `sig` is only set when there is an explicit
signature. The correct way to get the function signature is to call
`wasm_decl_get_signature`, which will check whether there is a func
type, and if not, will check for an explicit function signature.