Remove assert_{invalid,malformed} checks from wast2wast (#296)
Now that assert_invalid is checked by wasm-interp, all spec assertions
can be handled by the interpreter. Having a subset that are also handled
by wast2wasm is not useful.

Also, there was a bug in the option parser where passing an option that
didn't match could still work if a prefix was valid.