Use multiprocessing.pool (#312)
This replaces the ad-hoc mechanism we used before with threads. We'll
have to keep an eye on the Travis build, as we switched away from
multiprocessing before because it would flake occasionally. Hopefully
multiprocessing.pool doesn't have the same issue.

This also removes should_stop_on_error from the RunMultiThreaded path,
as it isn't allowed anyway (it early-outs when parsing arguments).