Fix crash using binary reader logging with error
When the binary logging is turned on, it can't use the
`reader->on_error` directly since that function expects that the
user_data is the BinaryReaderInterpreter context (for example), but it
is actually the LoggingContext.

To make it easier for users to use the default error handler, now they
return a bool saying whether the error was handled.

Some additional changes:

* Add a test that uses logging and succeeds
* Add a test that uses logging and has an error
* Change the `run-*-interp.py` scripts to use `-t` for tracing and `-v`
  for logging