Define the implicit func type before the function (#400)
When a function is defined without an explicit type, it will first
search the already defined (implicitly or otherwise) func types. If one
is found, it will use it. Otherwise it will define its own and use that.

The code previously would define this new type after the function, which
works but doesn't roundtrip properly because it ends up defining the
same func type twice.