Validate block signatures w/ multiple types
The parser and typechecker allow multiple types in block signatures:

    block i32 i32
      ...
    end

But these are not currently supported in the MVP, and there is no way to
serialize them. This change makes them a validation error.

This change also fixes the default location handling of non-terminal
nodes in the parser; they used to look weird because they would use the
ending column from the last symbol in the rule, even if that symbol was
on a different line.

Fixes #473.