Use templates instead of macros in interpreter (#564)
There's still more to be done, but this is a decent cleanup.

There is now a template type called `ValueTypeRep` which is the integer
representation of a given type. So for example, `ValueTypeRep<float>` =>
`uint32_t`.

Nearly all of the unary and binary operators are defined via template
functions that operate on these values.