Rename read_file -> ReadFile; now returns vector (#572)
This adds the `DataOrNull` function, which is used to get a std::vector
or std::string data(), but return `nullptr` if the container is empty. This
is needed because it is undefined behavior (at least according to ubsan)
if you call `data()` on an empty container.

It would be nicer to have the various read_* functions not take a void*,
size_t pair either, but that's work for another PR (perhaps introducing
a span-like API?)

I've also been considering having ReadFile return a wrapper object
instead, so this could just mmap the file.

Fixes #570.