Add tracing/logging output for atomic instructions
I forgot to add these in the last PR. It's easy to miss since there was
a default case on both switch statements.

This PR removes the default cases and explicitly enumerates the invalid
cases. It also adds two tests of all opcodes that can be
traced/disassembled in the interpreter.