[wasm-objdump] Continue reading sections on error (#636)
wasm-objdump wants to always try to read as much of the binary as
possible. This change adds a new BinaryReader option called
stop_on_first_error which can be set to false to enable this behavior.

Currently this will bail on any error in a section, but will try to read
the next section. We may expand the error-handling behavior to try and
recover inside of a section as well, but it's probably best to keep it
simple for now.

Because wasm-objdump does multiple passes over the binary, the errors
would normally be displayed multiple times. We also add a OnError
override which will suppress the error message for all passes other than
the first.