Support newer ubsan with -fno-sanitize-recover=all (#658)
UBSAN normally will only print errors and continue, but you can use
`-fno-sanitize-recover` to have it abort with an error message. At some
point the flag was made to require a parameter. We want to support both,
so we check at configuration time which one is allowed by the compiler.