Fix bad folding w/ imports and generating names
The name generation pass was adding an entry to the function binding
hash, but using the import index instead of the function/global/etc.
index. As a result, subsequent lookups would return the wrong function.

In this case, it would cause the wat folding code to use a function with
the wrong number of parameters.