Support different returncodes per command in tests
Each command can specify its expected returncode. If the returncode
doesn't match when the command is run, then no further commands are run
for that test.

This change also requires all ERROR directives to be specified after the
RUN commands, since they are added to a command in the same way as ARGS.