Store local types as Type+Count pairs
Since the binary format stores locals as Type+Count pairs, it is easy to
generate a function with a huge number of locals. The text format has
no way to compress this, so the resulting file will be huge.

However, if the binary file has an error, it would be useful to be able
to catch it without allocating a huge number of locals.

To do so, we store all locals as Type+Count pairs in the IR as well, and
provide accessor functions for getting the number of local types, the
type of a particular local index, etc.

This fixes issue #819.