Use `ReadCount` function in BinaryReader
See https://github.com/WebAssembly/wabt/issues/760 for more info.

We assume that a count precedes a list of items that are at least one
byte that are all contained in the same section. As a result, we can do
an up-front check to prevent over-allocation. This is the same technique
that the spec interpreter uses.