Better type-check errors locs @ end of block
Some type-checking only occurs at the end of a block, but the location
that is generated uses the last expression of that block. This can be
confusing if there are nested blocks, e.g.

```
1  block (result i32)
2    block
3      nop
4    end
5  end
```

This should produce an error at line 5, not line 3.

This PR stores the locations of the end of a block (either the `end` or
the closing parenthesis) so its easier to understand where the error
occurred.