[interp] Simplify host module imports (#889)
Importing from host modules used to be handled by a delegate, which is a
powerful enough solution to allow for all host import behaviors, but
makes the common case difficult and clunky.

It also assumed that every imported function/global/etc. needed to be
newly created and added to the environment, which is OK (though
wasteful) for functions, but is incorrect for mutable values like
memories and globals.

This change simplifies the common case by providing a set of functions
on `HostModule` that can be used to add exports, e.g.

    AppendFuncExport("add", {{Type::I32, Type::I32}, {Type::I32}},
                     AddFunc);

A generic interface for functions is provided as well, to allow for
automatically generating functions as needed. This is currently used for
functions like "host.print" in `wasm-interp`.