Add `dummy_workaround_...` function to fix Travis (#941)
Newer versions of emscripten will error out if you attempt to export
functions that aren't defined.

We added a dummy function to the export list to work around emscripten
issue 7073. It's fixed now, but we may want to keep this workaround for
a bit to allow for non-HEAD emscripten versions.

Fixes #939.