Fix wasm-objdump binary opcode output
Some instructions have very long encodings (longer than 9 bytes). For
those instructions, it is nice to write out all the bytes for that
instruction and their immediates. This used to be truncated, and
misaligned:

```
0002f1: fd 02 01 00 00 00 02 00 00 00 03 | v128.const 0x00000001 0x00000002 0x00000003 0x00000004
```

Now it is wrapped:

```
0002f0: fd 02 01 00 00 00 02 00 00 | v128.const 0x00000001 0x00000002 0x00000003 0x00000004
0002f9: 00 03 00 00 00 04 00 00 00 |
```

The instruction offset was also incorrect before for instructions with
a prefix, and is now fixed.