Rename duplicate locals
Similar to PR #512, duplicate locals are allowed in the `names` section,
but will produce an invalid wat file. We could support these properly
with custom annotations, but that isn't available yet. In the meantime,
we follow the same before as with functions and append `.1`, `.2`, etc
to the names to make them unique.