The great renaming
This huge PR does all the renaming as described in issue #933. It also
updates to the latest testsuite so the new names are used.

The old names of the MVP instructions are still supported for
convenience (though we should remove those too at some point), but the
old simd and atomic instruction names are no longer supported.