Update binary encoding of SIMD vector shuffle
The textual (WAT) representation of the vector shuffle immediate argument was
underspecified. Recently, https://github.com/WebAssembly/simd/pull/67 specified
it to be `i8x16.shuffle i5 i5 ... i5` (16 times `i5`), which is what all other
implementations were doing. WABT was reading the shuffle argument as 4 32-bit
hex integers that specified a `v128`, that is, `i8x16.shuffle 0x0 0x0 0x0 0x0`.

This PR updates WABT to conform to the modified draft of the SIMD extension.

Since there are more shuffle and permtue intrinsics in the WASM SIMD extension
pipeline, this PR adds a generic `iNxM` class implementing a packed array of `M`
`N`-bit wide integers.

The class can store the integer array in both textual and binary
representations. The textual repr is useful for WAT and type-checking, while the
binary repr is useful for the binary encoding.