Update testsuite; fix linking spec test (#1055)
When a module is instantiated, and the start function traps, the
contents of the memory and the table may have been modified. This case
is handled by the `assert_uninstantiable` check in a wast test.

In spectest-interp, assert_uninstantiable would instantiate the module,
but was incorrectly resetting the environment. In run-spec-wasm2c, the
`assert_uninstantiable` tests weren't being run at all. Now the module's
`init` function is run, and it is expected to trap.