Made common string lookups more robust.
Since the purpose of these strings is to display them to the user,
an "error string" is a more useful default than an assert, which
would require all callers to consistently check for the range
of valid values (which they didn't).

Tools like wasm-objdump should preferably be able to continue
outputting even in the face of illegal values.