Fix value dropping bug in return_call
The previous implementation was not properly dropping values on the
stack when performing a return_call or return_call_indirect.

In particular, the "keep count" (i.e. the number of values to keep that
are on the top of the stack) was set properly, but the "drop count" (the
number of values to drop below the keep count), was not.

This fixes issue #1108.