Fix floating point conversions on MSVC 32-bit
MSVC uses 53-bit precision for the x87 registers by default. In this
mode, it's not possible to properly round a large unsigned 64-bit
integer to a 32-bit float.

As a work-around, we can change the x87 floating-point precision, just
for these functions, and then reset it to the original precision.

A better solution may be to change the floating-point precision for the
entire program, though it's not clear what other effects that may have
on the generated code.

This change also re-enables the Appveyor 32-bit MSVC build.

Fixes issue #1113.