Visit catch even when its body is empty
The current visitor does not visit `catch` itself when its body is
empty. So even if the attached test with an empty `catch` body is
supposed to be a validation failure (because an `exnref` value is left
on the stack in the `catch` body), it passed the validator.

In more detail, in the current code,
Input: (should be a validation failure)
```
try
catch   ;; this pushes an exnref value onto the stack
end
```

Output: (wrong, `catch` is missing)
```
try
end
```

Now we correctly visit `catch` even when its body is empty, this test
correctly causes a validation failure.