wasm-decompile: Implement br/br_if
br becomes break or continue, and br_if the same, but prefixed
by an if(..).

These refer to the enclosing loop/block by generated label name,
this syntax may change.

We may later want to replace these by while/do-while/switch and
other special cases, but for now this is decently readable.

Also added ; to statements.