Initial commit of wasm-bugpoint.
wasm-bugpoint will be a delta debugger (see http://delta.tigris.org) for
WebAssembly, in the vein of bugpoint for llvm and bugpoint cranelift.

Presently in this initial commit, the basic framework of options processing,
loading a wasm file, writing it out, running the script and checking the scripts
return code are implemented. No bisection in this commit.