wasm-decompile: reworked how "stacky" code gets decompiled.
For example: multi-value, and void exps while there are non-void
exps on the stack.

It now uses temp variables instead of pseudo push/pop, as the latter
weren't particularly readable and had an ordering problem that was
hard to make intuitive.

The new system covers all possible situations, generates as few
variables as possible, has clearer comments, and tests.