reference-types: Improve reftype support (#1220)
Add Hostref type which is an internal subtype of anyref used to
refer to host objects.

Since anyref is an abstract base type it should be be stored directly
on the value stack.  The value stack instead must contain refs of
the concrete RefType enumeration (currently Null, Func, or Foreign).

These changes are needed for the implementation of the C API but are
separately useful for improving conformance with the reftype proposal.

Implement sub-typing rules for parameters and results when using
CallFunction in the interpreter.