Initial support for github actions
This adds a basic workflow that builds and tests wabt on all three
desktop platforms.

The plan is to extend this to completely replace travis and appveyor
in the future.

Remove the 2.7 requirement for python in CMakeLists.txt due to issue
with github actions where this ends up selected in the wrong (mingw)
version of python. See:
https://github.com/actions/setup-python/issues/40