reference-types: add final test: br_table
This requires some slighly different validation rules for MVP vs
interface-types.

There is a test in unreachable-invalid.wast that was removed in the
reference-types repo:
  type-br_table-label-num-vs-label-num-after-unreachable

This test depends on the old validation rules and still exists in the
master testsuite so we need to continue to support both sets of rules
for now.