wasm-decompile: use symbols from linking section for names.
This allows wasm .o files to have more readable names, or even final
linked modules if the linking information is preserved (with e.g.
--emit-relocs in LLD).

This is implemented as part of the WABT IR representation, so
benefits wasm2wat as well.

Named obtained this way are only set for functions if the function
doesn't also have a name in the name section, but is preferred over
the export name if there is one.