Improve help/usage message on invalid arguments
Don't print the full help message unless we actually ask for it,
but instead suggest `--help` as an option.

Also, print the program name before the error string.

Both these things seem match the GNU utilities I'm used to and I kind
like not seeing and entire screen of help text when I forget an
argument.